requirements.txt
aiogoogletrans==3.3.3
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.30.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cache2==0.2.2
filelock==3.16.1
flower==2.0.1
fsspec==2024.12.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.27.1
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kombu==5.4.2
langdetect==1.0.9
Mako==1.3.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pendulum==3.0.0
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==4.6.0
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.1
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
time-machine==2.16.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
